+++ kontrol-frontend/src/components/CytoscapeGraph/index.tsx
@@ -5,6 +5,9 @@ import stylesheet, { trafficNodeSelector } from "./stylesheet";
 import { useInterval } from "@react-hooks-library/core";
 import dagrePlugin, { dagreLayout } from "./plugins/dagre";
 import tippyPlugin, { createTooltip, TooltipInstance } from "./plugins/tippy";
+import { Flex } from "@chakra-ui/react";
+import Legend from "./Legend";
+import TooltipPortal from "./TooltipPortal";
 
 // register plugins with cytoscape
 cytoscape.use(dagrePlugin);
@@ -30,6 +33,11 @@ const CytoscapeGraph = ({
   // keep a ref to the cy instance. using state will cause infinite re-renders
   const cy = useRef<cytoscape.Core>();
   const tooltip = useRef<TooltipInstance | null>(null);
+  const [tooltipPortalElem, setTooltipPortalElem] =
+    useState<HTMLElement | null>(null);
+  const [hoveredNode, setHoveredNode] = useState<cytoscape.NodeSingular | null>(
+    null,
+  );
   const [animationsEnabled, setAnimationsEnabled] = useState(
     INIT_ANIMATIONS_ENABLED,
   );
@@ -52,12 +60,17 @@ const CytoscapeGraph = ({
         if (tooltip.current != null) {
           tooltip.current.destroy();
         }
-        tooltip.current = createTooltip(e.target);
+        const tooltipInstance = createTooltip(e.target);
+        if (tooltipInstance == null) return;
+        tooltip.current = tooltipInstance.instance;
+        setTooltipPortalElem(tooltipInstance.element);
+        setHoveredNode(e.target);
       });
       cy.current.on("mouseout", function () {
         if (tooltip.current != null) {
           tooltip.current.destroy();
         }
+        setHoveredNode(null);
       });
 
       // stop animations when the user is dragging nodes around
@@ -166,20 +179,22 @@ const CytoscapeGraph = ({
   );
 
   return (
-    <CytoscapeComponent
-      id="cytoscape-graph"
-      elements={elements}
-      style={{
-        width: "100%",
-        height: "100%",
-        minHeight: "267px",
-        display: "flex",
-      }}
-      layout={layout}
-      // @ts-expect-error cytoscape types are not great
-      stylesheet={stylesheet}
-      cy={handleCy}
-    />
+    <Flex w="100%" h="100%" minHeight="267px" position={"relative"}>
+      <Legend />
+      <TooltipPortal element={tooltipPortalElem} node={hoveredNode} />
+      <CytoscapeComponent
+        id="cytoscape-graph"
+        elements={elements}
+        style={{
+          width: "100%",
+          height: "100%",
+        }}
+        layout={layout}
+        // @ts-expect-error cytoscape types are not great
+        stylesheet={stylesheet}
+        cy={handleCy}
+      />
+    </Flex>
   );
 };